do postApplication v1

+++ deal/src/main/java/com/neoflex/creditconveyer/deal/domain/dto/LoanOfferDTO.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @Getter
 @Setter
 public class LoanOfferDTO   {
-    private BigDecimal applicationId;
+    private Long applicationId;
     private BigDecimal requestedAmount;
     private BigDecimal totalAmount;
     private Integer term;
